Crate jormungandrwallet

Modules

settings

time

Structs

EncryptingVoteKey

Error

Fragment

Proposal

Settings

SettingsRust

Wallet

Functions

iohk_jormungandr_delete_fragment⚠

delete the pointer

iohk_jormungandr_delete_spending_counters⚠

delete the inner buffer that was allocated by this library

iohk_jormungandr_fragment_from_raw⚠

deserialize a fragment from bytes

iohk_jormungandr_fragment_id⚠

get the ID of the provided fragment

iohk_jormungandr_symmetric_cipher_decrypt⚠

decrypt payload of the wallet transfer protocol

iohk_jormungandr_vote_proposal_new_private⚠

build the proposal object

iohk_jormungandr_vote_proposal_new_public⚠

build the proposal object

iohk_jormungandr_wallet_delete_buffer⚠

Delete a binary buffer that was returned by this library alongside with its length.

iohk_jormungandr_wallet_delete_error⚠

delete the pointer and free the allocated memory

iohk_jormungandr_wallet_delete_proposal⚠

delete the pointer

iohk_jormungandr_wallet_delete_settings⚠

delete the pointer and free the allocated memory

iohk_jormungandr_wallet_delete_string⚠

Delete a null terminated string that was allocated by this library

iohk_jormungandr_wallet_delete_wallet⚠

delete the pointer, zero all the keys and free the allocated memory

iohk_jormungandr_wallet_error_details⚠

Get a string describing the error, this will return an allocated null terminated string providing extra details regarding the source of the error.

iohk_jormungandr_wallet_error_to_string⚠

Get a string describing the error, this will return an allocated null terminated string describing the error.

iohk_jormungandr_wallet_id⚠

get the wallet id

iohk_jormungandr_wallet_import_keys⚠

recover a wallet from an account and a list of utxo keys

iohk_jormungandr_wallet_set_state⚠

update the wallet account state

iohk_jormungandr_wallet_spending_counters⚠

get the current spending counters for the (only) account in this wallet

iohk_jormungandr_wallet_total_value⚠

get the total value in the wallet

iohk_jormungandr_wallet_vote_cast⚠

build the vote cast transaction

Type Definitions

EncryptingVoteKeyPtr

ErrorPtr

FragmentPtr

ProposalPtr

SettingsPtr

WalletPtr